De-cruft the integration tests a small bit
